Deformations of curves in a K3
Finally understood in Algebraic Geometry notes why restricting a sheaf to a subscheme means tensoring by the structure sheaf. This should help me comput the restricted tangent cohomology and finally compute h^1(C), which is the dimension of the deformation space by Kodaira-Spencer. Of course it remains a mystery whether the dimension of the deformation space can indeed be interpreted as h^1(C)...